football-predictor/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==2.3.0
pandas==2.2.0
scipy==1.12.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.0
scikit-learn>=1.4.2
openpyxl>=3.1.2
pymc>=5.9.0
arviz>=0.17.1